move connection related files to "connections"
